requirements.txt
python-dotenv
openai
utils